+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 
+## [0.1.0-alpha.2] - 2021.04.02
+* add back button in conversation
+* add custom new Case Comment
 ## [0.1.0+alpha.1] - 2021.04.02
 * first alpha release
 ## [0.0.1] - 2021.04.01